Introduce TcpServer class
