+++ src/HttpServer.cpp
@@ -1,5 +1,6 @@
 /* vim: set ai et ts=4 sw=4: */
 
+#include <TcpServer.h>
 #include <HttpServer.h>
 #include <RegexCache.h>
 #include <Socket.h>
@@ -44,7 +45,7 @@ static void _httpNotFoundHandler(const HttpRequest&, HttpResponse& resp) {
     resp.setStatus(HTTP_STATUS_NOT_FOUND);
 }
 
-static void* _httpWorkerThreadProc(void* rawArg) {
+static void* HttpWorkerThreadProc(void* rawArg) {
     HttpWorkerThreadProcArg* arg = (HttpWorkerThreadProcArg*)rawArg;
     std::atomic_int& workersCounter = *arg->workersCounter;
 
@@ -211,24 +212,13 @@ void HttpWorker::run() {
  **********************************************************************
  */
 
-HttpServer::HttpServer()
-  : _listen_done(false)
-  , _listen_socket(-1)
-  , _workersCounter(0)
+HttpServer::HttpServer(): TcpServer(&HttpWorkerThreadProc)  
   , _handlers(nullptr) {
+
 }
 
+
 HttpServer::~HttpServer() {
-    // TODO: add timeout here and kill workers using pthread_kill if necessary
-
-    /* wait for running workers */
-    while(_workersCounter.load() > 0) {
-        usleep(10 * 1000); /* wait 10 ms */
-    }
-
-    if(_listen_done)
-        close(_listen_socket);
-
     while(_handlers != nullptr) {
         HttpHandlerListItem* next = _handlers->next;
         pcre_free(_handlers->regexp);
@@ -236,6 +226,7 @@ HttpServer::~HttpServer() {
         _handlers = next;
     }
 }
+
 
 void HttpServer::addHandler(HttpMethod method, const char* regexpStr, HttpRequestHandler handler) {
     const char* error;
@@ -257,98 +248,15 @@ void HttpServer::addHandler(HttpMethod method, const char* regexpStr, HttpReques
     _handlers = item;
 }
 
-/*
- * If client will close a connection send() will just return -1
- * instead of killing a process with SIGPIPE.
- */
-void HttpServer::_ignoreSigpipe() {
-    sigset_t msk;
-    sigemptyset(&msk);
-    sigaddset(&msk, SIGPIPE);
-    if(pthread_sigmask(SIG_BLOCK, &msk, nullptr) != 0)
-        throw std::runtime_error("HttpServer::_ignoreSigpipe() - pthread_sigmask() call failed");
-}
-
-void HttpServer::listen(const char* host, int port) {
-    struct sockaddr_in addr;
-
-    _ignoreSigpipe();
-
-    _listen_socket = socket(AF_INET, SOCK_STREAM, 0);
-    if(_listen_socket == -1)
-        throw std::runtime_error("HttpServer::listen() - socket() call failed");
-
-    addr.sin_family = AF_INET;
-    addr.sin_port = htons(port);
-    addr.sin_addr.s_addr = inet_addr(host);
-    if(bind(_listen_socket, (struct sockaddr*)&addr, sizeof(addr)) == -1)
-        throw std::runtime_error("HttpServer::listen() - bind() call failed");
-
-    if(::listen(_listen_socket, 100) == -1)
-        throw std::runtime_error("HttpServer::listen() - listen() call failed");
-
-    _listen_done = true;
-}
-
-/* One iteration of accept loop */
-void HttpServer::accept(const std::atomic_bool& terminate_flag) {
-    for(;;) {
-        struct timeval tv; /* we have to re-initialize these structures every time */
-        tv.tv_sec = 1L;
-        tv.tv_usec = 0;
-
-        fd_set rfds;
-        FD_ZERO(&rfds);
-        FD_SET(_listen_socket, &rfds);
-
-        int res = select(_listen_socket + 1, &rfds, (fd_set*)0, (fd_set*)0, &tv);
-        if(res > 0) /* ready to accept() */
-            break;
-
-        if(res < 0) { /* error */
-            if(errno == EINTR)
-                continue;
-            throw std::runtime_error("HttpServer::accept() - select() call failed: " + std::string(strerror(errno)));
-        }
-
-        /* timeout - check a terminate flag */
-        if(terminate_flag.load())
-            return;
-    }
-
-    int accepted_socket = ::accept(_listen_socket, 0, 0);
-    if(accepted_socket == -1)
-        throw std::runtime_error("HttpServer::accept() - accept() call failed");
-
-    // disable TCP Nagle's algorithm
-    int val = 1;
-    if(setsockopt(accepted_socket, IPPROTO_TCP, TCP_NODELAY, &val, sizeof(val)) < 0) {
-        close(accepted_socket);
-        throw std::runtime_error("HttpServer::accept() - setsockopt(2) error");
-    }
-
+void* HttpServer::createWTPArg(int accepted_socket, std::atomic_int* workersCounter){
     auto arg = new(std::nothrow) HttpWorkerThreadProcArg();
     if(arg == nullptr) {
         close(accepted_socket);
-        throw std::runtime_error("HttpServer::accept() - malloc() call failed");
+        throw std::runtime_error("HttpServer::createWTPArg() - malloc() call failed");
     }
-
     arg->socket = accepted_socket;
     arg->handlers = _handlers;
-    arg->workersCounter = &_workersCounter;
+    arg->workersCounter = workersCounter;
 
-    /*
-     * We need to increase workersCounter in the parent process to prevent race
-     * condition. Otherwise it's possible that parent process will terminate
-     * before child process increments a counter.
-     */
-    _workersCounter++;
-
-    pthread_t thr;
-    if(pthread_create(&thr, nullptr, _httpWorkerThreadProc, (void*)arg) != 0) {
-        _workersCounter--;
-        close(accepted_socket);
-        delete arg;
-        throw std::runtime_error("HttpServer::accept() - pthread_create() call failed");
-    }
+    return arg;
 }
